feat(mergeScan): add higher-order lettable version of mergeScan
