+++ src/operators/index.ts
@@ -41,6 +41,7 @@ export { merge } from './merge';
 export { mergeAll } from './mergeAll';
 export { mergeMap } from './mergeMap';
 export { mergeMapTo } from './mergeMapTo';
+export { mergeScan } from './mergeScan';
 export { min } from './min';
 export { multicast } from './multicast';
 export { observeOn } from './observeOn';
